refactor: address PR review — remove re-exports, fix config injection in format tables
- Remove `export { c, s, W, hr, effectiveWidth }` re-export from display.ts
- Update all callers (agent-controller, worker-controller, workspace-controller,
  settings-controller, input.ts, index.ts) to import c/s/W/hr from style.ts
- Fix clearBreak() to call effectiveWidth(W, getConfig().verbose) now that
  effectiveWidth no longer reads config directly
- Fix TOOL_RESULT_FMT._default and Bash entries to use { quiet, verbose }
  dispatch tables instead of inline getConfig().verbose calls
- Inject _thinkOutLoud from this.config into printBlock data so ASSISTANT_BLOCK_FMT
  thinking entry no longer calls getConfig() directly
- Update tests to import c/s/W/effectiveWidth from style.ts; update
  effectiveWidth tests to pass verbose param explicitly
