+++ src/agent/controllers/settings-controller.ts
@@ -1,4 +1,4 @@
-import { c } from "../views/display.js";
+import { c } from "../views/style.js";
 import type { WorkerDisplay } from "./worker-controller.js";
 import type { PickResult } from "../views/picker.js";
 import { Settings } from "../models/settings.js";
