+++ src/agent/index.ts
@@ -2,7 +2,8 @@ import "dotenv/config";
 import os from "node:os";
 import path from "node:path";
 import { fileURLToPath } from "url";
-import { Display, c, hr } from "./views/display.js";
+import { Display } from "./views/display.js";
+import { c, hr } from "./views/style.js";
 import { StatusBar } from "./views/status-bar.js";
 import { Input } from "./views/input.js";
 import { Picker } from "./views/picker.js";
