+++ tests/display.clear-break.test.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect } from "vitest";
-import { clearBreak, W } from "../src/agent/views/display.js";
+import { clearBreak } from "../src/agent/views/display.js";
+import { W } from "../src/agent/views/style.js";
 
 // Strip ANSI escape codes to get the visible text
 function strip(s: string) {
