+++ tests/repl.helpers.test.ts
@@ -7,7 +7,7 @@ import {
   fmtStats,
   fmtArgs,
 } from "../shared/formatters.js";
-import { c } from "../src/agent/views/display.js";
+import { c } from "../src/agent/views/style.js";
 
 describe("trunc", () => {
   it("returns short string unchanged", () => {
